_archiv/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.3
pydantic-settings==2.3.3
msal==1.28.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
cryptography==42.0.8
itsdangerous==2.2.0
celery==5.4.0
redis==5.0.6
python-multipart==0.0.9
pandas==2.2.2
